+++ packages/service-library/src/servicelib/fastapi/long_running_tasks/server.py
@@ -19,7 +19,6 @@ from ...long_running_tasks._task import (
 from ._dependencies import get_task_manager
 from ._server import setup
 
-
 __all__: tuple[str, ...] = (
     "get_task_manager",
     "setup",
